backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
distro==1.9.0
dotenv==0.9.9
eval_type_backport==0.3.1
faiss-cpu==1.13.2
fastapi==0.129.0
googleapis-common-protos==1.72.0
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
invoke==2.2.1
mistralai==1.12.4
numpy==2.4.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
protobuf==6.33.5
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
zipp==3.23.0
